==> gem_text/4_e2spo.ipynb <==
import pandas as pd
df_e = pd.read_csv('./data/3_e.csv')
df_e.info()
df_e.head()

entity_list = list(df_e['wiki_id'])
print(len(entity_list))
entity_list[:3]

from SPARQLWrapper import SPARQLWrapper, JSON
def query_subject(entity_id='Q43088'):
    # SPARQL endpoint for Wikidata
    sparql = SPARQLWrapper("https://query.wikidata.org/sparql")
    query = f"""
    SELECT ?subjectLabel ?predicateLabel ?objectLabel ?subject ?object 
    WHERE {{
      ?subject ?predicate ?object.
      ?subject rdfs:label ?subjectLabel.
      ?x wikibase:directClaim ?predicate.
      ?x rdfs:label ?predicateLabel.
      ?object rdfs:label ?objectLabel.
      BIND(wd:{entity_id} AS ?subject) 
      FILTER(LANG(?subjectLabel) = "en").
      FILTER(LANG(?predicateLabel) = "en").
      FILTER(LANG(?objectLabel) = "en").
    }}
    LIMIT 10000
    """
    # Set the query and return format
    sparql.setQuery(query)
    sparql.setReturnFormat(JSON)
    # Execute the query
    results = sparql.query().convert()
    return results

# query()

def query_object(entity_id='Q43088'):
    # SPARQL endpoint for Wikidata
    sparql = SPARQLWrapper("https://query.wikidata.org/sparql")
    query = f"""
    SELECT ?subjectLabel ?predicateLabel ?objectLabel ?subject ?object 
    WHERE {{
      ?subject ?predicate ?object.
      ?subject rdfs:label ?subjectLabel.
      ?x wikibase:directClaim ?predicate.
      ?x rdfs:label ?predicateLabel.
      ?object rdfs:label ?objectLabel.
      BIND(wd:{entity_id} AS ?object) 
      FILTER(LANG(?subjectLabel) = "en").
      FILTER(LANG(?predicateLabel) = "en").
      FILTER(LANG(?objectLabel) = "en").
    }}
    LIMIT 10000
    """
    # Set the query and return format
    sparql.setQuery(query)
    sparql.setReturnFormat(JSON)
    # Execute the query
    results = sparql.query().convert()
    return results
    
# query_object('Q43088')

import pandas as pd
def e2spo(entity_id='Q43088'):
    csv_data = []  # CSV header
    
    results = query_subject(entity_id)    
    for result in results["results"]["bindings"]:
        subject_label = result["subjectLabel"]["value"]
        predicate_label = result["predicateLabel"]["value"]
        object_label = result["objectLabel"]["value"]
        
        # Extract Wikidata IDs
        subject_id = result["subject"]["value"].split('/')[-1]  # Extract subject Wikidata ID
        object_id = result["object"]["value"].split('/')[-1]    # Extract object Wikidata ID
        
        # Add SPO triples to CSV data
        csv_data.append([subject_label, predicate_label, object_label, subject_id, object_id])

    results = query_object(entity_id)    
    for result in results["results"]["bindings"]:
        subject_label = result["subjectLabel"]["value"]
        predicate_label = result["predicateLabel"]["value"]
        object_label = result["objectLabel"]["value"]
        
        # Extract Wikidata IDs
        subject_id = result["subject"]["value"].split('/')[-1]  # Extract subject Wikidata ID
        object_id = result["object"]["value"].split('/')[-1]    # Extract object Wikidata ID
        
        # Add SPO triples to CSV data
        csv_data.append([subject_label, predicate_label, object_label, subject_id, object_id])
    
    df = pd.DataFrame(
        csv_data,
        columns=["subjectLabel","predicateLabel","objectLabel","subject","object"]
    )
    return df
    
df = e2spo(entity_id='Q43088')
df.info()
df.head()

df.tail()

```
from tqdm.auto import tqdm
list_df=[]
pbar = tqdm(entity_list)
for e in pbar:
    pbar.set_description(f'Query entity: {e:<15}')
    df=e2spo(entity_id = e)
    df['source']=e
    list_df.append(df)

# Concatenate the DataFrames row-wise (axis=0)
df = pd.concat(list_df, axis=0, ignore_index=True)
df.info()
df.head()
```

# pip install dask[dataframe] dask distributed "bokeh>=3.1.0"


from dask.distributed import Client, LocalCluster

# Set up a local Dask cluster
cluster = LocalCluster(n_workers=8)  # You can adjust settings here if needed (e.g., number of workers)
cluster

import dask.dataframe as dd
import pandas as pd
from dask import delayed
from tqdm.auto import tqdm
from dask.distributed import Client, LocalCluster

# Assume e2spo is a function that takes an entity ID and returns a Pandas DataFrame
# def e2spo(entity_id):
    # Your logic to fetch data goes here
    # Return a Pandas DataFrame
    # pass

# Create a delayed version of e2spo
@delayed
def delayed_e2spo(entity_id):
    df = e2spo(entity_id)
    df['source'] = entity_id
    return df

# Use context manager to create a Dask client
with Client(cluster) as client:
    
    # List to hold delayed tasks
    delayed_tasks = []
    
    # Use tqdm for progress bar
    pbar = tqdm(entity_list)
    for e in pbar:
        pbar.set_description(f'Query entity: {e:<15}')
        delayed_tasks.append(delayed_e2spo(e))
    
    # Compute the tasks in parallel using the Dask cluster
    list_df = delayed(delayed_tasks).compute()

    # Concatenate the DataFrames row-wise
    df = pd.concat(list_df, axis=0, ignore_index=True)

    # Convert Pandas DataFrame to Dask DataFrame for writing to Parquet
    ddf = dd.from_pandas(df, npartitions=1)

    # Write the Dask DataFrame to partitioned Parquet
    parquet_dir = './data/parquet_dir/'
    ddf.to_parquet(parquet_dir, partition_on=['source'], engine='pyarrow')

    # Display some information about the Dask DataFrame
    print("Partitioned Dask DataFrame written to Parquet.")


import pandas as pd

# Specify the directory where the partitioned Parquet files are stored
# parquet_dir = 'data/output_parquet/'

# Read the partitioned Parquet files into a Pandas DataFrame
df = pd.read_parquet(parquet_dir, engine='pyarrow')
df=df.reset_index(drop=True)
df.info()
df.head()


df_pivot = pd.pivot_table(
    df.reset_index(),
    # columns=['col1'],
    index=['predicateLabel'],
    values=['index'], 
    aggfunc={
        'index': ["count"],
    },
    fill_value=0
)
df_pivot.columns = ['_'.join(col).strip() for col in df_pivot.columns.values]
df_pivot.sort_values('index_count',ascending=False).head(50)

values_to_drop = {
    'diplomatic relation',
    'contains the administrative territorial entity',
    'emergency phone number',
    'topic\'s main Wikimedia portal',
    'public holiday',
    'highest judicial authority',
    'Wikidata property',
    'Wikimedia outline',
    'described by source',
    'different from',
    "topic's main category",
    'on focus list of Wikimedia project',
    "topic's main template",
    'maintained by WikiProject',
    'located in the administrative territorial entity',
    'top-level Internet domain',
    'flag',
    'currency',
    'language used'
}
df2 = df[~df['predicateLabel'].isin(values_to_drop)]
df_pivot = pd.pivot_table(
    df2.reset_index(),
    # columns=['col1'],
    index=['predicateLabel'],
    values=['index'], 
    aggfunc={
        'index': ["count"],
    },
    fill_value=0
)
df_pivot.columns = ['_'.join(col).strip() for col in df_pivot.columns.values]
df_pivot.sort_values('index_count',ascending=False).head(50)

df2.to_csv('./data/4_spo_e.csv')
df2.info()
df2.head()

import json
def spo2kg(df_spo_wiki):
    df = df_spo_wiki
    df_spo = df_spo_wiki[['subjectLabel','predicateLabel','objectLabel']]
    
    metadata = {}
    # Base Wikipedia URL for metadata
    wiki_base_url = "https://en.wikipedia.org/wiki/"
    
    for i in df.itertuples():
        # print(i.subjectLabel)
        
        # Add subject metadata (Wikidata ID and Wikipedia URL)
        if i.subjectLabel not in metadata:
            metadata[i.subjectLabel] = {
                "url": f"{wiki_base_url}{i.subjectLabel.replace(' ', '_')}",
                "wiki_id": i.subject
            }
        
        # Add object metadata (Wikidata ID and Wikipedia URL)
        if i.objectLabel not in metadata:
            metadata[i.objectLabel] = {
                "url": f"{wiki_base_url}{i.objectLabel.replace(' ', '_')}",
                "wiki_id": i.object
            }
    
    return df_spo, metadata

df_spo, metadata = spo2kg(df2)
df_spo.info()
df_spo.head()

csv_file= './data/4_data_e.csv'
metadata_file='./data/4_metadata_e.json'

df_spo.to_csv(csv_file, header=False, index=False)
with open(metadata_file, 'w', encoding='utf-8') as f:
    json.dump(metadata, f, indent=2)

!kg add -f ./data/4_data_e.csv
!kg meta -f ./data/4_metadata_e.json
!kg start
