==> gem_text/5_e2url.ipynb <==
from SPARQLWrapper import SPARQLWrapper, JSON

def get_wikiurl_wikidata(my_wiki_id ='Q83437'):
    sparql = SPARQLWrapper("https://query.wikidata.org/sparql")

    query = """
    SELECT ?entity ?entityLabel ?wikipedia_url WHERE {
      BIND(wd:[[my_wiki_id]] AS ?entity).
      ?wikipedia_url schema:about ?entity;
                     schema:isPartOf <https://en.wikipedia.org/>.
      SERVICE wikibase:label { bd:serviceParam wikibase:language "en". }
    }
    """.replace('[[my_wiki_id]]',my_wiki_id)
    
    # Set up the SPARQL query
    sparql.setQuery(query)
    sparql.setReturnFormat(JSON)
    
    # Execute the query and retrieve the results
    results = sparql.query().convert()
    
    # Process and print results
    for result in results["results"]["bindings"]:
        entity_label = result["entityLabel"]["value"]
        wikipedia_url = result["wikipedia_url"]["value"]
        if wikipedia_url is not None:
            return wikipedia_url
    raise Exception('wikipedia_url is not found')

get_wikiurl_wikidata()

import pandas as pd
df_e=pd.read_csv('./data/3_e.csv')
df_e.info()
df_e.head()

pip install monadsquishy -U

from monadsquishy import Squishy, sf
sq_config = {
    'transformations':[
        {
            'input_table': df_e,
            'transformed_path':'./staging/url',
            'exploded_path':'./staging/url',
            'out_columns': {
                 'wiki_id': {
                     'input':'wiki_id',
                     'funcs':[lambda x:x, ],
                 },
                 'url': {
                     'input':'wiki_id',
                     'funcs':[get_wikiurl_wikidata, ],
                 }
             }
        }
    ]
}
sq=Squishy(sq_config)
sq.run()

sq.clean_report()

sq.dirty_report()

sq.output()

entity_list = list(df_e['wiki_id'])[:3]
len(entity_list)
entity_list

from tqdm.auto import tqdm
urls={}
pbar = tqdm(entity_list)
for e in pbar:
    pbar.set_description(f'Query entity: {e:<15}')
    urls[e] = get_wikiurl_wikidata(e)

urls
