==> gem_text/41_spo2e.ipynb <==
import pandas as pd
csv_file= './data/4_data_e.csv'
metadata_file='./data/4_metadata_e.json'

df=pd.read_csv(csv_file, header=None)
df.columns=['subjectLabel','predicateLabel','objectLabel']
df.info()
df.head()

df_pivot = pd.pivot_table(
    df.reset_index(),
    # columns=['col1'],
    index=['predicateLabel'],
    values=['index'], 
    aggfunc={
        'index': ["count"],
    },
    fill_value=0
)
df_pivot.columns = ['_'.join(col).strip() for col in df_pivot.columns.values]
df_pivot.sort_values('index_count',ascending=False).head(20)

values_to_drop = {
    'main subject',
    'category combines topics',
    'expressed in',
}
df2 = df[~df['predicateLabel'].isin(values_to_drop)]
df2.info()
df2.head()

df3 = df2[['subjectLabel','objectLabel']].melt()[['value']].drop_duplicates()
df3.columns = ['wiki_id']
df3.info()
df3.head()

df3.to_csv('./data/41_e.csv')
